+++ test/wave/WaveFileMock.h
@@ -16,7 +16,7 @@ public:
 	WaveFileMock();
 	~WaveFileMock();
 	MOCK_METHOD0(getNumberOfSamples, unsigned int());
-	MOCK_METHOD1(getSample, int(int i));
+	MOCK_METHOD1(getSample, double(unsigned int i));
 	MOCK_METHOD0(getBitsPerSample, unsigned int());
 
 };
